+++ frontend/src/components/Portfolio/Sections/Positions.js
@@ -1,39 +1,16 @@
-import React, { useState, useEffect } from 'react';
+import { React } from 'react';
 import Container from 'react-bootstrap/Container';
 import Table from 'react-bootstrap/Table';
 import '../../../styles/Portfolio/Positions.scss';
 
 const Positions = (props) => {
-    // const { username } = props;
-
-    const [league, setLeague] = useState();
-    const [portfolio, setPortfolio] = useState();
-
-    const getPortfolio = async () => {
-        const response = await fetch(`${process.env.REACT_APP_API_URL}/league/portfolio`, {
-            method: 'GET',
-            headers: {
-                Authorization: `Bearer ${localStorage.getItem('token')}`,
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-                leagueName: league,
-            }),
-        });
-        setPortfolio(await response.json());
-    };
-
-    useEffect(() => {
-        console.log(props);
-        setLeague('Test League 4');
-        getPortfolio();
-    });
+    const { portfolio } = props;
 
     return (
         <div>
             <Container className="portfolio-positions-container">
                 {portfolio && <h1>{portfolio.cashAvailable}</h1>}
-                <Table striped bordered className="positions-list">
+                <Table striped bordered hover variant="dark" className="positions-list">
                     <thead>
                         <th>Symbol</th>
                         <th>Description</th>
@@ -43,6 +20,23 @@ const Positions = (props) => {
                         <th>Average Cost Basis</th>
                         <th>Total Change</th>
                     </thead>
+                    {portfolio
+                        && (
+                            <tbody>
+                                {portfolio.holdings
+                                && Object.keys(portfolio.holdings).map((ticker) => (
+                                    <tr>
+                                        <td>{ticker}</td>
+                                        <td>{portfolio.holdings[ticker].equityName}</td>
+                                        <td>{portfolio.holdings[ticker].quantity}</td>
+                                        <td>{portfolio.holdings[ticker].currentPrice}</td>
+                                        <td>{portfolio.holdings[ticker].totalValue}</td>
+                                        <td>{portfolio.holdings[ticker].costBasis}</td>
+                                        <td>{portfolio.holdings[ticker].totalChange}</td>
+                                    </tr>
+                                ))}
+                            </tbody>
+                        )}
                 </Table>
             </Container>
         </div>
